+++ src/sam/macros/index.js
@@ -53,6 +53,7 @@ async function exec (arc, cloudformation, stage, options) {
   let transforms = arc.macros || []
   // Always run the following internal macros:
   transforms.push('legacy-api') // Use legacy REST APIs instead of HTTP APIs for @http; must run before other macros
+  transforms.push('http-ver')   // Reconfigure HTTP APIs to use legacy REST API payload format
   transforms.push('api-path')   // Updates @cdn, @ws stage URL paths
   transforms.push('arc-env')    // Gets and sets env vars for functions
   transforms.push('static')     // Sets SPA, S3 prefix, etc. in root handler